chore(resources): move create resource from middleware to hooks
