+++ integration/fixtures/ruleset.rest.yaml
@@ -4,72 +4,97 @@ rules:
       methods: ["GET", "POST"]
       target: "http://127.0.0.1:13777/"
       frontends:
-      - name: some_rest_1
-        path: "/basic"
-        method: "GET"
-        hooks:
-        - name: authz
-          config:
-            action: authz_action
-      - name: some_rest_2
-        path: "/basic-authn"
-        method: "GET"
-        middlewares:
-        - name: prefix
-          config:
-            strip: "/basic-authn"
-        - name: basic_auth
-          config:
-            users:
-              - user: user
-                # password must be hashed using MD5, SHA1, or BCrypt(recommended) using htpasswd
-                password: $apr1$RfxoV6GP$.GsGgD580H5FOuUfTzKZh0 # md5: password
-        hooks:
-          - name: authz
-            config:
-              action: authz_action
-      - name: some_rest_3
-        path: "/basic-authn-bcrypt"
-        method: "GET"
-        middlewares:
-          - name: prefix
-            config:
-              strip: "/basic-authn-bcrypt"
-          - name: basic_auth
-            config:
-              users:
-                - user: user
-                  # password must be hashed using MD5, SHA1, or BCrypt(recommended) using htpasswd
-                  password: $2y$10$F814ZwQPt8VHYIayIqeEReSeZz8dDCNX93/rKI82SqJu9I2Bn6Hau # BCrypt: password
-        hooks:
-          - name: authz
-            config:
-              action: authz_action
-      - name: some_rest_4
-        path: "/basic-authz"
-        method: "POST"
-        middlewares:
-          - name: prefix
-            config:
-              strip: "/basic-authz"
-          - name: basic_auth
-            config:
-              users:
-                - user: user
-                  password: $apr1$RfxoV6GP$.GsGgD580H5FOuUfTzKZh0
-                  capabilities: [ "ping:foo", "do:bar" ]
-              scope:
-                action: "ping:{{.project}}"
-                attributes:
-                  project:
-                    type: json_payload
-                    key: project
-        hooks:
-          - name: authz
-            config:
-              action: authz_action
-              attributes:
-                project:
-                  key: project
-                  type: json_payload
-                  source: request
+        - name: some_rest_1
+          path: "/basic"
+          method: "GET"
+        #        hooks:
+        #        - name: authz
+        #          config:
+        #            action: authz_action
+        - name: some_rest_2
+          path: "/basic-authn"
+          method: "GET"
+          middlewares:
+            - name: prefix
+              config:
+                strip: "/basic-authn"
+            - name: basic_auth
+              config:
+                users:
+                  - user: user
+                    # password must be hashed using MD5, SHA1, or BCrypt(recommended) using htpasswd
+                    password: $apr1$RfxoV6GP$.GsGgD580H5FOuUfTzKZh0 # md5: password
+        #        hooks:
+        #          - name: authz
+        #            config:
+        #              action: authz_action
+        - name: some_rest_3
+          path: "/basic-authn-bcrypt"
+          method: "GET"
+          middlewares:
+            - name: prefix
+              config:
+                strip: "/basic-authn-bcrypt"
+            - name: basic_auth
+              config:
+                users:
+                  - user: user
+                    # password must be hashed using MD5, SHA1, or BCrypt(recommended) using htpasswd
+                    password: $2y$10$F814ZwQPt8VHYIayIqeEReSeZz8dDCNX93/rKI82SqJu9I2Bn6Hau # BCrypt: password
+        #        hooks:
+        #          - name: authz
+        #            config:
+        #              action: authz_action
+        - name: some_rest_4
+          path: "/basic-authz"
+          method: "POST"
+          middlewares:
+            - name: prefix
+              config:
+                strip: "/basic-authz"
+            - name: basic_auth
+              config:
+                users:
+                  - user: user
+                    password: $apr1$RfxoV6GP$.GsGgD580H5FOuUfTzKZh0
+                    capabilities: [ "ping:foo", "do:bar" ]
+                scope:
+                  action: "ping:{{.project}}"
+                  attributes:
+                    project:
+                      type: json_payload
+                      key: project
+      #        hooks:
+      #          - name: authz
+      #            config:
+      #              action: authz_action
+      #              attributes:
+      #                project:
+      #                  key: project
+      #                  type: json_payload
+      #                  source: request
+    - name: dagger
+      path: /daggers
+      target: "http://127.0.0.1:3000/"
+      methods: [ "GET", "POST" ]
+      frontends:
+        - name: create dagger
+          path: "/daggers"
+          method: "POST"
+#          hooks:
+#            - name: authz
+#              config:
+#                attributes:
+#                  project:
+#                    key: project
+#                    type: json_payload
+#                  organization:
+#                    key: organization
+#                    type: json_payload
+#                  team:
+#                    key: team
+#                    type: json_payload
+#                  resource:
+#                    key: urns.#.id
+#                    type: json_payload
+
