+++ middleware/authz/authz.go
@@ -1,14 +1,11 @@
 package authz
 
 import (
-	"context"
 	"fmt"
 	"net/http"
 	"strings"
 
 	"github.com/odpf/shield/api/handler"
-	"github.com/odpf/shield/model"
-
 	"github.com/odpf/shield/middleware"
 	"github.com/odpf/shield/pkg/body_extractor"
 	"github.com/odpf/shield/structs"
@@ -165,93 +162,8 @@ func (c *Authz) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 	}
 
 	// use permissionAttributes & config.Action here
-	resources, err := createResources(permissionAttributes)
-	if err != nil {
-		c.log.Error(err.Error())
-		return
-	}
-	for _, resource := range resources {
-		res, err := c.Deps.V1beta1.ResourceService.Create(context.Background(), resource)
-		if err != nil {
-			c.log.Error(err.Error())
-			return
-		}
-		c.log.Info(fmt.Sprintf("Resource %s created", res.Id))
-	}
 
 	c.next.ServeHTTP(rw, req)
-}
-
-func createResources(permissionAttributes map[string]interface{}) ([]model.Resource, error) {
-	var resources []model.Resource
-	projects, err := getAttributesValues(permissionAttributes["project"])
-	if err != nil {
-		return nil, err
-	}
-
-	orgs, err := getAttributesValues(permissionAttributes["organization"])
-	if err != nil {
-		return nil, err
-	}
-
-	teams, err := getAttributesValues(permissionAttributes["team"])
-	if err != nil {
-		return nil, err
-	}
-
-	resourceList, err := getAttributesValues(permissionAttributes["resource"])
-	if err != nil {
-		return nil, err
-	}
-
-	namespace, err := getAttributesValues(permissionAttributes["namespace"])
-	if err != nil {
-		return nil, err
-	}
-
-	if len(projects) < 1 || len(orgs) < 1 || len(teams) < 1 || len(resourceList) < 1 {
-		return nil, fmt.Errorf("projects, organizations, resource, and team are required")
-	}
-
-	for _, org := range orgs {
-		for _, project := range projects {
-			for _, team := range teams {
-				for _, res := range resourceList {
-					resources = append(resources, model.Resource{
-						Name:           res,
-						OrganizationId: org,
-						ProjectId:      project,
-						GroupId:        team,
-						NamespaceId:    namespace[0],
-					})
-				}
-			}
-		}
-	}
-	return resources, nil
-}
-
-func getAttributesValues(attributes interface{}) ([]string, error) {
-	var values []string
-	switch attributes.(type) {
-	case []string:
-		for _, i := range attributes.([]string) {
-			values = append(values, i)
-		}
-	case string:
-		values = append(values, attributes.(string))
-	case []interface{}:
-		for _, i := range attributes.([]interface{}) {
-			values = append(values, i.(string))
-		}
-	case interface{}:
-		values = append(values, attributes.(string))
-	case nil:
-		return values, nil
-	default:
-		return values, fmt.Errorf("unsuported attribute type %v", attributes)
-	}
-	return values, nil
 }
 
 func (w Authz) notAllowed(rw http.ResponseWriter) {
